debug(Storybook/UI): move Context from preview to the Story itself
